prozy lambda response & status code

+++ build.sh
@@ -6,6 +6,6 @@ rm README.md
 mv new_README.md README.md
 m2r README.md
 mv README.rst docs/README.rst
-rm -r dist
+[ -d dist ] && rm -r dist
 poetry build
 twine check dist/*
